# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: WhatTools API
  version: "1.0.0"
  description: B2B SaaS platform providing professional-grade tools for Whatnot sellers.
  contact:
    name: WhatTools
    url: https://whattools.trector.com
servers:
  - url: http://localhost:5000/api/v1
    description: Local development
  - url: https://whattools.trector.com/api/v1
    description: Production

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        data: { type: "null" }
        meta: { type: object }
        errors:
          type: array
          items:
            type: object
            properties:
              code: { type: string }
              message: { type: string }
    RegisterRequest:
      type: object
      required: [email, password, account_name]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        account_name: { type: string }
        name: { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string }
    LogoutRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string }
    InventoryItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        category_id: { type: string, format: uuid, nullable: true }
        cogs: { type: number }
        quantity: { type: integer }
        status: { type: string, enum: [available, sold, reserved, listed] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateItemRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string }
        category_id: { type: string, format: uuid }
        cogs: { type: number, default: 0 }
        quantity: { type: integer, default: 1 }
        status: { type: string, enum: [available, sold, reserved, listed] }
    Category:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        created_at: { type: string, format: date-time }
    CreateCategoryRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string }
    InviteRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
        role: { type: string, enum: [admin, member], default: member }

paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new account and user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "201":
          description: Account created
        "400":
          description: Validation error
        "409":
          description: Email already exists

  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Login successful, tokens returned
        "401":
          description: Invalid credentials
        "403":
          description: Account locked

  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshRequest" }
      responses:
        "200":
          description: New token pair
        "401":
          description: Invalid or expired refresh token

  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke refresh token
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LogoutRequest" }
      responses:
        "200":
          description: Token revoked

  /auth/password-reset:
    post:
      tags: [Auth]
      summary: Request password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "200":
          description: Reset email sent (always returns 200)

  /auth/password-reset/confirm:
    post:
      tags: [Auth]
      summary: Confirm password reset with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token: { type: string }
                password: { type: string, minLength: 8 }
      responses:
        "200":
          description: Password updated
        "400":
          description: Invalid or expired token

  /auth/verify-email:
    post:
      tags: [Auth]
      summary: Verify email address with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        "200":
          description: Email verified

  /users/me:
    get:
      tags: [Profile]
      summary: Get current user profile
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: User profile
    put:
      tags: [Profile]
      summary: Update current user profile
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                email: { type: string, format: email }
      responses:
        "200":
          description: Profile updated
    patch:
      tags: [Profile]
      summary: Change password
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: { type: string }
                new_password: { type: string, minLength: 8 }
      responses:
        "200":
          description: Password changed

  /account:
    get:
      tags: [Account]
      summary: Get account details
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: Account info
    put:
      tags: [Account]
      summary: Update account
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
      responses:
        "200":
          description: Account updated
    delete:
      tags: [Account]
      summary: Delete account (owner only)
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: Account deleted

  /account/members:
    get:
      tags: [Account]
      summary: List team members
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: List of team members

  /account/invite:
    post:
      tags: [Account]
      summary: Invite a team member
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InviteRequest" }
      responses:
        "201":
          description: Invitation sent
        "403":
          description: Permission denied
        "409":
          description: User already a member

  /account/members/{user_id}/role:
    put:
      tags: [Account]
      summary: Update member role
      security: [{ BearerAuth: [] }]
      parameters:
        - name: user_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { type: string, enum: [owner, admin, member] }
      responses:
        "200":
          description: Role updated

  /account/members/{user_id}:
    delete:
      tags: [Account]
      summary: Remove a team member
      security: [{ BearerAuth: [] }]
      parameters:
        - name: user_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Member removed

  /items:
    get:
      tags: [Inventory]
      summary: List inventory items
      security: [{ BearerAuth: [] }]
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string, enum: [available, sold, reserved, listed] }
        - name: category_id
          in: query
          schema: { type: string, format: uuid }
        - name: cursor
          in: query
          schema: { type: string, format: uuid }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        "200":
          description: Paginated list of items
    post:
      tags: [Inventory]
      summary: Create inventory item
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateItemRequest" }
      responses:
        "201":
          description: Item created
        "400":
          description: Validation error
        "403":
          description: Tier limit reached

  /items/deleted:
    get:
      tags: [Inventory]
      summary: List soft-deleted items (trash)
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: Deleted items within 30-day retention

  /items/{item_id}:
    get:
      tags: [Inventory]
      summary: Get single item
      security: [{ BearerAuth: [] }]
      parameters:
        - name: item_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Item details
        "404":
          description: Item not found
    put:
      tags: [Inventory]
      summary: Update item
      security: [{ BearerAuth: [] }]
      parameters:
        - name: item_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateItemRequest" }
      responses:
        "200":
          description: Item updated
    delete:
      tags: [Inventory]
      summary: Soft-delete item
      security: [{ BearerAuth: [] }]
      parameters:
        - name: item_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Item soft-deleted

  /items/{item_id}/restore:
    post:
      tags: [Inventory]
      summary: Restore soft-deleted item
      security: [{ BearerAuth: [] }]
      parameters:
        - name: item_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Item restored
        "403":
          description: Tier limit would be exceeded
        "404":
          description: Deleted item not found

  /categories:
    get:
      tags: [Inventory]
      summary: List categories
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: List of categories
    post:
      tags: [Inventory]
      summary: Create category
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCategoryRequest" }
      responses:
        "201":
          description: Category created

  /categories/{category_id}:
    get:
      tags: [Inventory]
      summary: Get single category
      security: [{ BearerAuth: [] }]
      parameters:
        - name: category_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Category details
    put:
      tags: [Inventory]
      summary: Update category
      security: [{ BearerAuth: [] }]
      parameters:
        - name: category_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Category updated
    delete:
      tags: [Inventory]
      summary: Delete category
      security: [{ BearerAuth: [] }]
      parameters:
        - name: category_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Category deleted

  /csv/upload:
    post:
      tags: [CSV Import]
      summary: Upload CSV file for import
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Upload successful, returns job with headers and preview

  /csv/{job_id}:
    get:
      tags: [CSV Import]
      summary: Get import job status
      security: [{ BearerAuth: [] }]
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Job details with progress

  /csv/{job_id}/map:
    post:
      tags: [CSV Import]
      summary: Submit column mapping and start import
      security: [{ BearerAuth: [] }]
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [column_mapping]
              properties:
                column_mapping:
                  type: object
                  additionalProperties: { type: string }
      responses:
        "200":
          description: Import completed with results

  /health:
    get:
      tags: [Health]
      summary: Aggregated health check
      responses:
        "200":
          description: All services healthy
        "503":
          description: One or more services unhealthy